Move MotorManager to its own file
